// src/api/generated/model/analysisJobStatus.ts
/**
 * Generated by orval v7.0.0 🍺
 * Do not edit manually.
 * MHWs Talisman Vision API
 * Backend API and Vision inference engine for MHWs Talisman Vision Module.
 * OpenAPI spec version: 0.1.0
 */

export type AnalysisJobStatus = typeof AnalysisJobStatus[keyof typeof AnalysisJobStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const AnalysisJobStatus = {
  pending: 'pending',
  processing: 'processing',
  completed: 'completed',
  failed: 'failed',
  cancelled: 'cancelled',
} as const;