// src/api/generated/model/validationStatus.ts
/**
 * Generated by orval v7.0.0 🍺
 * Do not edit manually.
 * MHWs Talisman Vision API
 * Backend API and Vision inference engine for MHWs Talisman Vision Module.
 * OpenAPI spec version: 0.1.0
 */

export type ValidationStatus = typeof ValidationStatus[keyof typeof ValidationStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ValidationStatus = {
  processing: 'processing',
  valid: 'valid',
  needs_selection: 'needs_selection',
  error: 'error',
} as const;